fixed bugs
