+++ src/utils/constant.ts
@@ -1,0 +1,11 @@
+export const AVAILABLE_COLORS = [
+  { value: "WHITE", label: "ขาว", code: "#ffffff" },
+  { value: "PURPLE", label: "ม่วง", code: "#7758a9" },
+  { value: "RED", label: "แดง", code: "#aa2639" },
+  { value: "DARK_RED", label: "แดงคริสมาสต์", code: "#c30000" },
+  { value: "ORANGE", label: "ส้ม", code: "#F48c08" },
+  { value: "YELLO_LEMON", label: "เหลืองมะนาว", code: "#e6e007" },
+  { value: "DARK_GREEN", label: "เขียวแก่", code: "#2f5e1e" },
+  { value: "GREEN_APPLE", label: "เขียวแอปเปิ้ล", code: "#80c06d" },
+  { value: "BLUE", label: "ฟ้า", code: "#5997bc" },
+];
